requirements.txt
pytesseract
pillow
opencv-python
numpy
transformers
torch
sacremoses